Add support for marshalling dict-like objects.
The `is_dict_like` function in `schema.py` was just checking if the
object's type was a dict.  This causes it to treat any dict subclass as
a model, which fails because dict subclasses are not models.

Updating to check if the item is an instance of `collections.Mapping`
(or `collections.abc.Mapping` on python 3) fixes this.