Add new 'use_spec_url_for_base_path' bool option to spec.CONFIG_DEFAULTS
This new option addresses issues arising from 'basePath' not being
present in a given spec.

According to the 2.0 spec, "If it is not included, the API is served directly
under the host. The value MUST start with a leading slash (/)".

10836363 made the SwaggerClient compliant with the spec (basePath is now
implicitly set to '/' if field is missing from spec). However, this broke
existing code that utilizes the SwaggerClient (see Issue #299).

The current (spec-compliant) behaviour remains with this new option as
it defaults to False. Overriding this default with 'True' means the path
element of the URL used to retrieve the spec will be used instead.